General settings interface renamed according to requirement
