+++ frontend/src/app/core/models/IGeneralAvailabilitySettings.ts
