+++ frontend/src/app/modules/availability/new-slot/general/general.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { IGeneralAvailabilitySettings } from '@core/models/generalAvailabilitySettings';
+import { IGeneralAvailabilitySettings } from '@core/models/IGeneralAvailabilitySettings';
 
 @Component({
     selector: 'app-general',
